requirements.txt
asgiref==3.2.3
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.1.0
django-reversion==3.0.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
et-xmlfile==1.0.1
freetype-py==2.1.0.post1
future==0.17.1
httplib2==0.9.2
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mysqlclient==1.4.6
numpy==1.15.4
odfpy==1.3.6
opencv-python==4.2.0.32
openpyxl==2.5.11
pandas==1.0.3
Pillow==5.3.0
pipenv==2018.11.26
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2018.7
PyYAML==3.13
requests==2.20.1
scipy==1.1.0
six==1.11.0
sqlparse==0.3.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.7.9
virtualenv-clone==0.5.3
xlrd==1.1.0
xlwt==1.3.0
